switched to concern instead of base class
