+++ app/components/op_turbo_sandbox/projects/index_component.rb
@@ -28,7 +28,9 @@
 
 module OpTurboSandbox
   module Projects
-    class IndexComponent < OpTurbo::Component
+    class IndexComponent < ApplicationComponent
+      include OpTurbo::Streamable
+
       def initialize(projects:)
         super()
         @projects = projects
